refactor(auth): apply review round on block_on, shared formatter, and docs
Drop the redundant async block around with_authentication (block_on
takes the impl Future directly); merge the two expiry formatters into
core::utils::format_expiry_utc; trim the strict-discovery and keyring
dependency comments; move the cred_env tests to a sibling file per the
project convention.
