+++ core/Cargo.toml
@@ -83,12 +83,8 @@ icu_properties = "2.1.2"
 # backend reports "absent" and callers fall back to `SYSAND_CRED_*`
 # (section 9 taxonomy). On musl the keyring feature still compiles; the
 # backend reports "absent".
-# On non-musl targets keyring 4.x's `v1` feature selects the native
-# backend per OS: Keychain on macOS, Credential Manager on Windows, and
-# the pure-Rust zbus Secret Service backend on glibc Linux. zbus needs no
-# system libdbus (nothing to vendor) and its crypto-rust session keeps
-# openssl out of the tree, so glibc builds (including CI lint runs with
-# --all-features) need no system libraries.
+# keyring's `v1` feature selects the native backend per OS: Keychain on macOS, Credential
+# Manager on Windows, and the pure-Rust (i.e. no libdbus or openssl) zbus Secret Service on Linux
 [target.'cfg(not(all(target_os = "linux", target_env = "musl")))'.dependencies]
 keyring = { version = "4.1.5", optional = true, default-features = false, features = ["v1"] }
